Changing long to double.

+++ src/main/java/tr/edu/ku/ulgen/dto/UlgenDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class UlgenDto {
     private Long userId;
-    private Long[] location;
+    private Double[] location;
     private Integer activeUser;
     private String userCity;
 }
